Split out managing_step_by_step_pages_spec.rb into multiple features
This file had grown to test a number of features. I've now extracted out the
index page features (view and filtering) into one file, and the creation of
step by steps into another file, leaving the original file to be solely about
the management of an existing step by step.

As `then_I_see_the_new_step_by_step_page` is used in multiple places, I've
moved it out to a shared place.
